requirements.txt
appnope==0.1.0
backcall==0.1.0
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
Click==7.0
coverage==4.5.4
decorator==4.3.0
Faker==0.9.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
ipython==7.0.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
MarkupSafe==1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
python-dateutil==2.7.3
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.12
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
